#557 - mu: put core streams I/O in streams
