+++ src/mu/core/lib.rs
@@ -11,13 +11,12 @@ use {
             types::Tag,
         },
         features::feature::{Core as _, Feature},
-        streams::write::Core as _,
+        streams::{core::StreamBuilder, write::Core as _},
         types::{
-            core_stream::Stream,
             fixnum::{Core as _, Fixnum},
             function::Function,
             namespace::Namespace,
-            stream::StreamBuilder,
+            stream::Stream,
             symbol::{Core as _, Symbol},
             vector::Vector,
             vector_image::Core as _,
