+++ src/mu/core/quasi.rs
@@ -14,8 +14,8 @@ use crate::{
     streams::read::Core as _,
     types::{
         cons::{Cons, Core as _},
-        core_stream::{Core as _, Stream},
         namespace::Namespace,
+        stream::{Core as _, Stream},
         symbol::{Core as _, Symbol},
     },
 };
