+++ src/mu/streams/mod.rs
@@ -2,6 +2,7 @@
 //  SPDX-License-Identifier: MIT
 
 //! system streams
+pub mod core;
 pub mod operator;
 pub mod read;
 pub mod system;
