+++ src/mu/streams/write.rs
@@ -13,11 +13,11 @@ use crate::{
     types::{
         char::{Char, Core as _},
         cons::{Cons, Core as _},
-        core_stream::{Core as _, Stream},
         fixnum::{Core as _, Fixnum},
         float::{Core as _, Float},
         function::{Core as _, Function},
         namespace::{Core as _, Namespace},
+        stream::{Core as _, Stream},
         struct_::{Core as _, Struct},
         symbol::{Core as _, Symbol},
         vector::Vector,
