+++ src/mu/types/mod.rs
@@ -1,10 +1,9 @@
 //  SPDX-FileCopyrightText: Copyright 2022 James M. Putnam (putnamjm.design@gmail.com)
 //  SPDX-License-Identifier: MIT
 
-//! env types module
+//! types
 pub mod char;
 pub mod cons;
-pub mod core_stream;
 pub mod fixnum;
 pub mod float;
 pub mod function;
